Store user configuration in the users home directory.
Creepy is currently designed to run from the source directory and
save its plugin configuration in the current directory.  When
installed by the Debian Package to /usr/share the plugins do not work
because creepy does not have the required priviledges there.

This patch reads the configuration from /usr/share and then when
writing the new configuration, writes the non-default values to
~/.creepy.  Thereafter the configuration is read from both /usr/share
and ~/.creepy. This keeps the original configuration file intact and
allow default values to change when the package is upgraded without
having to edit files in the users home directories.